Feat: improve path, refactor emitter
