+++ src/index.js
@@ -3,7 +3,7 @@ import { homedir } from "node:os";
 import { parse, path, user } from "./utils/index.js";
 import { commandLine } from "./commandLine.js";
 import { emitter } from "./emitter.js";
-import { youAreHere } from "./utils/path.js";
+import library from "./library/index.js";
 
 const FileManager = {
   pathToWorkingDirectory: homedir(),
@@ -17,10 +17,10 @@ path.youAreHere(FileManager.pathToWorkingDirectory);
 commandLine.on("line", async (line) => {
   try {
     const { module, command, lineArguments } = parse.toParameters(line);
-    await emitter.execute(module, command, lineArguments, FileManager);
+    await library[module][command](lineArguments, FileManager);
   } catch (error) {
     emitter.throw(error);
   } finally {
-    youAreHere(FileManager.pathToWorkingDirectory);
+    path.youAreHere(FileManager.pathToWorkingDirectory);
   }
 });
